Compute stats across all requests, not just the first
